Refactor test item pipeline into modular package
