+++ library/testitem_pipeline/__init__.py
@@ -1,0 +1,137 @@
+"""Test item pipeline helpers exposed as a public API."""
+
+from __future__ import annotations
+
+import json
+import requests
+
+from .catalog import (
+    LoadMoleculeHierarchyLookup,
+    PARENT_LOOKUP_SOURCE_CACHE,
+    PARENT_LOOKUP_SOURCE_LOOKUP,
+    PARENT_LOOKUP_SOURCE_PARTIAL,
+    PARENT_LOOKUP_SOURCE_SKIPPED,
+    PARENT_LOOKUP_SOURCE_SYNC,
+    ParentEnrichmentPreparation,
+    ParentEnrichmentResult,
+    ParentLookupPreparedData,
+    ParentLookupStats,
+    _DEFAULT_CATALOG_CFG,
+    _MOLECULE_HIERARCHY_COLUMNS,
+    _TYPO_PARENT_COLUMN,
+    _merge_parent_stats,
+    attach_parent_molecule_ids,
+    ensure_no_parant_column,
+    load_molecule_hierarchy_lookup,
+    load_parent_catalog,
+    molecule_catalog,
+    prepare_parent_enrichment,
+    query_parent_catalog,
+    run_parent_enrichment,
+    update_parent_catalog_cache,
+    write_parent_catalog_cache,
+)
+from .cli import (
+    ReadInputIdsResult,
+    TestitemFetchError,
+    TestitemPipelineOptions,
+    TestitemPipelineStageError,
+    _FETCH_ERROR_SAMPLE_SIZE,
+    _prepare_pubchem_api_cfg,
+    apply_testitem_enrichment,
+    fetch_testitems,
+    finalize_output,
+    integrate_missing_identifiers,
+    read_input_ids,
+    run_testitem_pipeline,
+)
+from .pubchem import (
+    PUBCHEM_CID_CACHE_ENCODING,
+    PUBCHEM_COLUMNS,
+    _CID_CACHE_MISSING,
+    _PUBCHEM_CACHE_SCHEMA_VERSION,
+    _PUBCHEM_SESSION_LOCK,
+    _PUBCHEM_SESSION_SIGNATURE,
+    _load_pubchem_cid_cache,
+    _prepare_pubchem_caches,
+    _prefetch_parents,
+    _pubchem_session_signature,
+    _resolve_pubchem_cids,
+    _write_pubchem_cid_cache,
+    add_pubchem_data,
+    augment_pubchem,
+    resolve_pubchem_cid,
+)
+from library import pubchem_library as pl
+from library import testitem_enrichment
+from library.chembl_client import ChemblClient
+from library.clients import pubchem as pc
+from library.log import logger
+from library.metadata import file_sha256, write_meta_yaml
+from library.table_quality import analyze_table_quality
+
+__all__ = [
+    "ReadInputIdsResult",
+    "TestitemPipelineOptions",
+    "TestitemFetchError",
+    "TestitemPipelineStageError",
+    "_DEFAULT_CATALOG_CFG",
+    "_FETCH_ERROR_SAMPLE_SIZE",
+    "_MOLECULE_HIERARCHY_COLUMNS",
+    "_PUBCHEM_CACHE_SCHEMA_VERSION",
+    "_PUBCHEM_SESSION_LOCK",
+    "_PUBCHEM_SESSION_SIGNATURE",
+    "_TYPO_PARENT_COLUMN",
+    "_CID_CACHE_MISSING",
+    "_merge_parent_stats",
+    "_prepare_pubchem_api_cfg",
+    "_prepare_pubchem_caches",
+    "_prefetch_parents",
+    "_pubchem_session_signature",
+    "_resolve_pubchem_cids",
+    "_merge_pubchem_properties",
+    "_load_pubchem_cid_cache",
+    "_write_pubchem_cid_cache",
+    "ChemblClient",
+    "analyze_table_quality",
+    "LoadMoleculeHierarchyLookup",
+    "ParentEnrichmentPreparation",
+    "ParentEnrichmentResult",
+    "ParentLookupPreparedData",
+    "ParentLookupStats",
+    "PARENT_LOOKUP_SOURCE_CACHE",
+    "PARENT_LOOKUP_SOURCE_LOOKUP",
+    "PARENT_LOOKUP_SOURCE_PARTIAL",
+    "PARENT_LOOKUP_SOURCE_SKIPPED",
+    "PARENT_LOOKUP_SOURCE_SYNC",
+    "attach_parent_molecule_ids",
+    "ensure_no_parant_column",
+    "load_molecule_hierarchy_lookup",
+    "load_parent_catalog",
+    "molecule_catalog",
+    "prepare_parent_enrichment",
+    "query_parent_catalog",
+    "run_parent_enrichment",
+    "update_parent_catalog_cache",
+    "write_parent_catalog_cache",
+    "resolve_pubchem_cid",
+    "add_pubchem_data",
+    "augment_pubchem",
+    "apply_testitem_enrichment",
+    "read_input_ids",
+    "integrate_missing_identifiers",
+    "fetch_testitems",
+    "run_testitem_pipeline",
+    "finalize_output",
+    "PUBCHEM_COLUMNS",
+    "PUBCHEM_CID_CACHE_ENCODING",
+    "analyze_table_quality",
+    "file_sha256",
+    "json",
+    "logger",
+    "pc",
+    "pl",
+    "requests",
+    "testitem_enrichment",
+    "write_meta_yaml",
+]
